Update to MetaXR v63

+++ MRUtilityKit/Source/MRUtilityKit/Public/MRUtilityKitBlobShadowComponent.h
@@ -69,7 +69,7 @@ public:
 
 	void BeginPlay() override;
 	void TickComponent(float DeltaTime, enum ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
-	void ComputeOwner2DBounds(FVector& Origin, FVector2D& Extent, float& Yaw) const;
+	void ComputeOwner2DBounds(FVector& Origin, FVector2D& Extent, double& Yaw) const;
 
 protected:
 	UPROPERTY()
